refactor(utils): rename digest
